+++ src/main/resources/db/migration/V1.0.1__create_table_login.sql
@@ -4,7 +4,7 @@ CREATE TABLE Login(
 			primary key,
 	username varchar(256),
 	password varchar(256),
-	roles_id integer,
+	roles_id bigserial,
 	customer_id bigserial,
 	seller_company_id bigserial,
 	shipper_id bigserial
