+++ src/main/resources/db/migration/V1.1.4__create_table_payment.sql
@@ -5,7 +5,7 @@ CREATE TABLE Payment(
       payment_amount integer,
       payment_state integer,
       timestamp timestamp,
-      order_id integer,
-      cc_id integer,
+      order_id bigserial,
+      cc_id bigserial,
       payment_status integer  -- 0 не оплачен 1 оплаченный
 );
