+++ src/main/resources/db/migration/V1.1.7__alter_table_products.sql
@@ -1,6 +1,6 @@
 ALTER TABLE Products
-ADD COLUMN seller_company_id integer;
+ADD COLUMN seller_company_id bigserial;
 
 ALTER TABLE Products
-ADD COLUMN special_characteristic_id integer;
+ADD COLUMN special_characteristic_id bigserial;
 );
